pulse/requirements/base.txt
# PULSE Base Requirements
# Tested and compatible versions

streamlit==1.39.0
pandas==2.2.3
duckdb==1.1.3
azure-kusto-data==4.6.2
azure-identity==1.19.0
openai==1.58.1
python-dotenv==1.0.1
pyyaml==6.0.2
plotly==5.24.1
